+++ src/controllers/client.ts
@@ -2,8 +2,8 @@ import Express from "express";
 import { google } from "googleapis";
 import { Client } from "../models/client";
 import { check_dup_client } from "../utils/clientDup";
+import { CLIENT_SHEET_ID } from "../utils/sheetID";
 import { serialize_rows, sheet_auth } from "../utils/sheets";
-import { CLIENT_SHEET_ID } from "./checkins";
 
 export const read_one_client: Express.RequestHandler = async (req, res) => {
     const sheet = google.sheets("v4");
