+++ src/utils/sheetID.ts
@@ -1,0 +1,3 @@
+export const EVENT_SHEET_ID = "EVENTS";
+export const CHECKIN_SHEET_ID = "CHECKINS";
+export const CLIENT_SHEET_ID = "CLIENT";
